split secondary types
