+++ src/Poseidon/CLI/Rectify.hs
@@ -4,28 +4,27 @@ module Poseidon.CLI.Rectify (
     runRectify, RectifyOptions (..), PackageVersionUpdate (..), ChecksumsToRectify (..)
     ) where
 
-import           Poseidon.EntityTypes    (renderNameWithVersion)
-import           Poseidon.GenotypeData   (GenotypeDataSpec (..))
-import           Poseidon.Package        (PackageReadOptions (..),
-                                          PoseidonPackage (..),
-                                          defaultPackageReadOptions,
-                                          readPoseidonPackageCollection,
-                                          writePoseidonPackage)
-import           Poseidon.SecondaryTypes (ContributorSpec (..),
-                                          VersionComponent (..),
-                                          updateThreeComponentVersion)
-import           Poseidon.Utils          (PoseidonIO, getChecksum, logDebug,
-                                          logInfo)
+import           Poseidon.Contributor   (ContributorSpec (..))
+import           Poseidon.EntityTypes   (renderNameWithVersion)
+import           Poseidon.GenotypeData  (GenotypeDataSpec (..))
+import           Poseidon.Package       (PackageReadOptions (..),
+                                         PoseidonPackage (..),
+                                         defaultPackageReadOptions,
+                                         readPoseidonPackageCollection,
+                                         writePoseidonPackage)
+import           Poseidon.Utils         (PoseidonIO, getChecksum, logDebug,
+                                         logInfo)
+import           Poseidon.Version       (VersionComponent (..),
+                                         updateThreeComponentVersion)
 
-import           Control.DeepSeq         ((<$!!>))
-import           Control.Monad.IO.Class  (liftIO)
-import           Data.List               (nub)
-import           Data.Maybe              (fromJust)
-import           Data.Time               (UTCTime (..), getCurrentTime)
-import           Data.Version            (Version (..), makeVersion,
-                                          showVersion)
-import           System.Directory        (doesFileExist, removeFile)
-import           System.FilePath         ((</>))
+import           Control.DeepSeq        ((<$!!>))
+import           Control.Monad.IO.Class (liftIO)
+import           Data.List              (nub)
+import           Data.Maybe             (fromJust)
+import           Data.Time              (UTCTime (..), getCurrentTime)
+import           Data.Version           (Version (..), makeVersion, showVersion)
+import           System.Directory       (doesFileExist, removeFile)
+import           System.FilePath        ((</>))
 
 data RectifyOptions = RectifyOptions
     { _rectifyBaseDirs              :: [FilePath]
