+++ src/Poseidon/Chronicle.hs
@@ -2,33 +2,32 @@
 
 module Poseidon.Chronicle where
 
-import           Poseidon.EntityTypes    (HasNameAndVersion (..))
-import           Poseidon.Package        (PoseidonPackage (..))
-import           Poseidon.SecondaryTypes (VersionComponent (..),
-                                          updateThreeComponentVersion)
-import           Poseidon.Utils          (Env (..), PoseidonException (..),
-                                          PoseidonIO, TestMode (..))
+import           Poseidon.EntityTypes   (HasNameAndVersion (..))
+import           Poseidon.Package       (PoseidonPackage (..))
+import           Poseidon.Utils         (Env (..), PoseidonException (..),
+                                         PoseidonIO, TestMode (..))
+import           Poseidon.Version       (VersionComponent (..),
+                                         updateThreeComponentVersion)
 
-import           Control.Monad.Catch     (throwM)
-import           Control.Monad.IO.Class  (liftIO)
-import           Control.Monad.Reader    (asks)
-import           Data.Aeson              (FromJSON, ToJSON, object, parseJSON,
-                                          toJSON, withObject, (.!=), (.:),
-                                          (.:?), (.=))
-import qualified Data.ByteString         as B
-import           Data.Function           (on)
-import           Data.List               (elemIndex)
-import           Data.Maybe              (fromMaybe)
-import qualified Data.Set                as S
-import           Data.Time               (Day, UTCTime (..), getCurrentTime)
-import           Data.Version            (Version, makeVersion)
-import           Data.Yaml               (decodeEither')
-import           Data.Yaml.Pretty        (defConfig, encodePretty,
-                                          setConfCompare, setConfDropNull)
-import           GitHash                 (getGitInfo, giHash)
-import           System.Directory        (createDirectoryIfMissing,
-                                          makeAbsolute)
-import           System.FilePath         (makeRelative, takeDirectory)
+import           Control.Monad.Catch    (throwM)
+import           Control.Monad.IO.Class (liftIO)
+import           Control.Monad.Reader   (asks)
+import           Data.Aeson             (FromJSON, ToJSON, object, parseJSON,
+                                         toJSON, withObject, (.!=), (.:), (.:?),
+                                         (.=))
+import qualified Data.ByteString        as B
+import           Data.Function          (on)
+import           Data.List              (elemIndex)
+import           Data.Maybe             (fromMaybe)
+import qualified Data.Set               as S
+import           Data.Time              (Day, UTCTime (..), getCurrentTime)
+import           Data.Version           (Version, makeVersion)
+import           Data.Yaml              (decodeEither')
+import           Data.Yaml.Pretty       (defConfig, encodePretty,
+                                         setConfCompare, setConfDropNull)
+import           GitHash                (getGitInfo, giHash)
+import           System.Directory       (createDirectoryIfMissing, makeAbsolute)
+import           System.FilePath        (makeRelative, takeDirectory)
 
 data PoseidonPackageChronicle = PoseidonPackageChronicle
     { snapYamlTitle            :: String
